+++ helpers.php
@@ -8,5 +8,5 @@ defined('\\ABSPATH') || exit;
 
 function get_icon(string $name): SvgIcon
 {
-    return new SvgIcon((file_get_contents(PLUGIN_PATH . "assets/icons/$name.svg")));
+    return new SvgIcon((file_get_contents(PLUGIN_PATH . "icons/$name.svg")));
 }
