+++ simple-analytics.php
@@ -32,9 +32,9 @@ require __DIR__ . '/src/Setting.php';
 require __DIR__ . '/src/SettingName.php';
 require __DIR__ . '/src/TrackingPolicy.php';
 require __DIR__ . '/src/ScriptCollection.php';
-require __DIR__ . '/src/Scripts/Script.php';
-require __DIR__ . '/src/Scripts/HideScriptId.php';
-require __DIR__ . '/src/Scripts/HasAttributes.php';
+require __DIR__ . '/src/Scripts/Contracts/Script.php';
+require __DIR__ . '/src/Scripts/Contracts/HideScriptId.php';
+require __DIR__ . '/src/Scripts/Contracts/HasAttributes.php';
 require __DIR__ . '/src/Scripts/AnalyticsScript.php';
 require __DIR__ . '/src/Scripts/AutomatedEventsScript.php';
 require __DIR__ . '/src/Scripts/InactiveScript.php';
