+++ src/Scripts/Contracts/HideScriptId.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace SimpleAnalytics\Scripts\Contracts;
+
+interface HideScriptId
+{
+}
