--- src/Scripts/HideScriptId.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace SimpleAnalytics\Scripts;
-
-interface HideScriptId
-{
-}
