+++ src/Scripts/InactiveScript.php
@@ -2,6 +2,7 @@
 
 namespace SimpleAnalytics\Scripts;
 
+use SimpleAnalytics\Scripts\Contracts\Script;
 use const SimpleAnalytics\PLUGIN_URL;
 
 class InactiveScript implements Script
@@ -9,7 +10,7 @@ class InactiveScript implements Script
     #[\Override]
     public function path(): string
     {
-        return PLUGIN_URL . 'assets/js/inactive.js';
+        return PLUGIN_URL . 'js/inactive.js';
     }
 
     #[\Override]
